backend/functions/requirements.txt
functions-framework==3.*
pyyaml==6.0
google-cloud-storage==2.7.0
google-cloud-documentai==2.6.0
pandas==1.5.3
numpy==1.24.3
